# docs/swagger.yaml
basePath: /
definitions:
  model.Apikey:
    properties:
      createdAt:
        type: string
      createdBy:
        type: string
      data:
        type: string
      deletedAt:
        type: string
      description:
        type: string
      hint:
        type: string
      id:
        type: string
      organizationId:
        type: string
      platformId:
        type: string
      type:
        type: string
      updatedAt:
        type: string
    type: object
  model.Contract:
    properties:
      address:
        type: string
      createdAt:
        type: string
      deactivatedAt:
        type: string
      deletedAt:
        type: string
      functions:
        items:
          type: string
        type: array
      id:
        type: string
      name:
        type: string
      networkId:
        type: string
      platformId:
        type: string
      updatedAt:
        type: string
    type: object
...
  model.RequestContractCreate:
    properties:
      address:
        type: string
      functions:
        items:
          type: string
        type: array
      name:
        type: string
      networkId:
        type: string
      platformId:
        type: string
    type: object
  model.RequestContractUpdate:
    properties:
      address:
        type: string
      functions:
        items:
          type: string
        type: array
      name:
        type: string
      networkId:
        type: string
    type: object
...
host: platform-api.string-api.xyz
info:
  contact:
    email: support@stringxyz.com
    name: Platform Management API Support
    url: http://string.xyz
  description: Platform Management API for managing Organizations, Platforms, API
    keys, etc.
  title: Platform Management API
  version: "1.0"
paths:
  /apikeys:
    get:
      consumes:
      - application/json
      description: Get all API keys
      parameters:
      - description: Platform ID
        in: query
        name: platformId
        type: string
      - description: Limit the number of returned results
        in: query
        name: limit
        type: integer
      - description: Offset the starting point of returned results
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Apikey'
            type: array
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Get all API keys
      tags:
      - Apikey
    post:
      consumes:
      - application/json
      description: Create an API key
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Type of API key (public or secret)
        in: query
        name: type
        type: string
      - description: Platform ID
        in: query
        name: platformId
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Apikey'
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Create API key
      tags:
      - Apikey
  /apikeys/{id}:
    delete:
      consumes:
      - application/json
      description: Delete API key by ID
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Delete API key
      tags:
      - Apikey
    get:
      consumes:
      - application/json
      description: Get API key by ID
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Apikey'
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Get API key
      tags:
      - Apikey
    put:
      consumes:
      - application/json
      description: Update API key by ID
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      - description: API key description
        in: body
        name: description
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Apikey'
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Update API key
      tags:
      - Apikey
  /contracts:
    get:
      consumes:
      - application/json
      parameters:
      - description: Platform ID
        in: query
        name: platformId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Contract'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Get all contracts
      tags:
      - Contract
    post:
      consumes:
      - application/json
      parameters:
      - description: Contract Create Request
        in: body
        name: RequestContractCreate
        required: true
        schema:
          $ref: '#/definitions/model.RequestContractCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Contract'
        "400":
          description: Bad Request
          schema:
            type: integer
        "409":
          description: Conflict
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Create a new contract
      tags:
      - Contract
  /contracts/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Contract ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Contract'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Get a specific contract
      tags:
      - Contract
    patch:
      consumes:
      - application/json
      parameters:
      - description: Contract ID
        in: path
        name: id
        required: true
        type: string
      - description: Contract Update Request
        in: body
        name: RequestContractUpdate
        required: true
        schema:
          $ref: '#/definitions/model.RequestContractUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Contract'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Update a contract
      tags:
      - Contract
  /contracts/{id}/deactivate:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Contract ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Contract'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Deactivate a contract
      tags:
      - Contract
  /contracts/{id}/reactivate:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Contract ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Contract'
        "400":
          description: Bad Request
          schema:
            type: integer
        "500":
          description: Internal Server Error
          schema:
            type: integer
      security:
      - JWTAuth: []
      - BearerAuth: []
      summary: Reactivate a contract
      tags:
      - Contract
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    in: header
    name: StringAdminJWT
  RefreshBearerAuth:
    type: apiKey
    in: header
    name: StringAdminRefreshToken
  JWTAuth:
    type: apiKey
    in: cookie
    name: StringAdminJWT
  RefreshTokenAuth:
    type: apiKey
    in: cookie
    name: StringAdminRefreshToken
swagger: "2.0"